[UPDATE] Refactor examination follow commitment
